mirroring along axis
this patch changes the behavior of mirroring (drill holes and back side)
to mirror on the coordinate axis and not on the center axis of the
board.

mirroring based on board bounding box can be troublesome when ngc files
are to be merged with processing results of other gerber files which
have different bounding boxes.