slimmer gtk initialization
initializing gtk is necessary to avoid segfaults like this:

Failed to wrap object of type 'GdkPixbuf'. Hint: this error is commonly
caused by failing to call a library init() function.

but the old initialization depended on a working X display to be
present.

contary to old comment, the line had nothing to do with command line
parsing.