better fix for bug 3089930
the --drill option starts exactly like --drill-size, which confuses the
boost option parser if it is configured to guess (which is the default
for whatever reasons).

the new parsing style now explicitly disables guessing.

(and now that the sequence of options can be arbitrary again, shoving
them in an appropriate sequence again)