a different approach to the basename setting
this has the same purpose, but another method than bkubicek's --basename
patch from
0004-milling-of-holes-using-a-small-mill-head-activated.patch

as the current patch queue already contains infrastructure to set the
individual output file names, this version just modifies those arguments
and leaves the rest of the code unchanged.

the methods used in this patch are none that i like, but i didn't find
a better way to feed additional arguments to the vm.