configure G64 tolerance depending on dpi
the G64 command defines how far off the tool position may be from the
defined path. as a pixel is the limit of precision in the calculations,
it makes no sense to have more precision in cutting.

the value of 0.75 (actually less than a pixel) was chosen by
experimenting with the emc2 software demo machine -- it's the smallest
value that produced cuts that didn't show steps from aliasing.