dwell to avoid takeoff/landing errors
with smoothing enabled, every time the tool plunges in to start cutting,
the Z down movement is combined with an XY movement, resulting in a
landing (or, later, takeoff) arc that causes some copper not to be
removed.

putting dwell points with zero time at appropriate places (after pluges,
before retracts) solves that problem. the dwell point caused as a side
effect before going to the first position should not cause any trouble
(it does show at the coordinate origin, which could potentially confuse
users).

while this problem became more grave with the latest patch for
dpi-dependent G64, it was present before as well.