implemented outline width compensation
the ability to have outlines as lines with a given width got lost when
contour masking was added. this adds a new implementation that is
compatible with masking.

unlike the old version, this uses component growing to grow the lines
back. while this does not give ultimate rendering results, it is
required to get precise masking (so the previews are accurate and the
engraving bit will remove the copper neatly at the border of where the
milling head will later cut off the board).