Rework outline filling with vectorial core
With the old raster core the outline was filled by filling the closed
lines with the same color of the outline line, creating a filled area
larger than the intended one (caused by the line thickness), requiring
an additional correction based on the value of outline-width.

With the vectorial core this correction is no longer needed, as the
software can just skip the buffering of the closed lines and keep them
as rings (instead of converting them to polygons).
Therefore outline-width is now ignored with the vectorial core.