fix mirror-absolute
this changes the mirror axis if mirror-absolute and zero-start are both
active. The previous behavior was inconsitent with the manual. The new
behavior allows the user to zero a back layer on the gerber zero,
because it is consistent on front and back layer. This fix changes only
vectorial behavior.