Preserve thermal reliefs by default when voronoi is true
When voronoi mode is enabled, thermal reliefs are detected by looking for holes in traces that have no other traces inside of them.  When holes are found, they are added to the path to be drilled.

preserve_thermal_reliefs is enabled by default and only has an effect when voronoi is enabled.