Use a better bridge-finding algorithm.
Previously, the algorithm would attempt to put bridges on the longest
path segments.  The new algorithm looks at all posisble segments that
can fit a bridge and tries to find a subset that will maximize the
distance between the two closest bridges.  This is a hard problem so
it uses a hueristic to find a local best.  In practice, this is
usually a good selection.