Limit spike length to the nearest trace.
This is just in case the buffer operation creates a sharp corner that
leads to a very long spike, for example, if a very sharp corner is
created from a narrow trapezoid that is bg::buffered.  All spikes are
limited to not cross any other trace that needs to be drawn.