Unify tool output between milling and drilling
In #236 the following script is proposed to split the mill files into one for each tool (cutter diameter):
https://github.com/pcb2gcode/pcb2gcode/files/2914413/split-gcode.txt

This script does not work as-is for drill files to split them into individual files for differently sized drills. The reason is that tool size messages are formatted differently, for example:
```(MSG, Change tool bit to drill size 0.4 mm)```
vs:
```(MSG, Change tool bit to mill diameter 1.20000mm)```

Notice the additional space in the drill tool change message.

This patch removes the space between diameter and unit for drill size output, making the script usable for both drill and mill files.