Add a test for overlapping bridges
It's possible for a bridge to be added that changes the length of the
segment on which a different bridge will be placed.  So instead of
just storing the pointer to the segment, we should be storing the
pointer to the segment and also the length from the start to the
center of the bridge.