Fix autoleveller interpolation bug
If the autoleveller wants to calculate the height of a point that is
directly between two measured points, it will interpolate between just
those two points, instead of using all four points in the adjacent
square. This fixes a bug where, if the point being calculated was at the
far left or top of the design, the resulting G code would ask for points
off the edge of the design that hadn't been measured.