Make autoleveller use generated variable numbers
The autoleveller's constructor generates a collection of unique variable
numbers that it can use when calculating its interpolated heights,
however these where only used in one place (when moving the tool
vertically down at the start of a line) and variables 1, 2 and 3 were
used in the rest of the code. This caused the output G code to reference
the wrong variable when calculating the milling depth at the start of a
line. The autoleveller now uses its generated variables throughout its
code.