fix: Use a spike mask instead of line intersections.
Checking for intersections fails sometimes when the milled line
matches the voronoi boundary because the spike doesn't have an inner
path against which to make an intersection.  This can cause the spike
to dig into the trace a little.

Rather than check if a spike intersects a previous milling line, just
mask the spikes by the area of the previous milling.  This should do
the same work, more or less, but without the pathological case when
the previous milling paths match the current milling path at some or
all of the path.