fix: Close unclosed contours
According to the [Gerber Format Specification](https://www.bosco.co.za/DownloadDocs/The_Gerber_Format_Specification.pdf) section 4.6.1, a contour may only be ended by a D02 or G37 when the contour is closed.  For example, a rectangle must have four line segments, not three.

EasyEDA seems to be making shapes missing a line segment and assumes that interpreters will close the contour.  `pcb2gcode` can support this by checking that the region is closed.  If not, add a point at the end which matches the first point so that the contour is closed.

This fixes #577