fix: Use spikes_keep_in for regular spikes, too.
The spike math might lead a spike to point away from the trace instead
of toward it.  This is unlikely but when it happens, it might aim at a
neighboring trace!  And we don't yet have the neighboring trace's keep
out so we can't use it.

Rather than defer the computation of spikes, which would be difficult,
we will just keep the spikes within the voronoi region of the trace.

This only applies to regular traces, not to thermal reliefs nor
voronoi milling.

This fixes #636