test: Remove parts of the project-controller for performance.
The project-controller takes too long too run and only a small section
of it is needed for testing.  Remove parts of the gerber file so that
it will run faster.  We must be careful to not remove the parts that
are fixed in 2453cbd30500048779062d627c64a8cf3cf5cbe9 .