fix: Always perform at least one step.
Laser cutters might need to set zwork to 0 or stepsize to 0, which
caused the code to compute that no milling/cutting needs to be done at
all!  It seems unlikely that someone would not want to do any work at
all so we'll make sure that there is always at least one pass.